+++ playground/CHANGELOG.md
@@ -1,6 +1,14 @@
 # playground
 
+## 0.0.164-canary.18
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@8.0.0-canary.18
+
 ## 0.0.164-canary.17
+
 ## 0.0.172
 
 ### Patch Changes